Tweak using PortableClipboard for dialog textboxes
Subclassing Form reduces the duplication of boilerplate code.
