+++ SIL.Windows.Forms/SIL.Windows.Forms.csproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="12.0">
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -577,6 +577,7 @@
     </Compile>
     <Compile Include="ImageToolbox\ImageAcquisitionService.cs" />
     <Compile Include="Miscellaneous\PortableClipboard.cs" />
+    <Compile Include="Miscellaneous\FormUsingPortableClipboard.cs" />
     <Compile Include="Miscellaneous\IUserInterfaceMemory.cs" />
     <Compile Include="Miscellaneous\SILAboutBox.cs">
       <SubType>Form</SubType>
